Update Hale.Core for xplat
