+++ Backend/Hale.Core/Handlers/AgentDistHandler.Core.cs
@@ -1,0 +1,26 @@
+﻿#if NETCOREAPP2_1
+using Hale.Core.Data.Entities.Nodes;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Hale.Core.Handlers
+{
+    internal partial class AgentDistHandler
+    {
+
+        private void GenerateTemplateVersionFile(string templateVersionPath)
+        {
+            throw new NotImplementedException();
+        }
+
+
+        private void CreatePackage(Node host, bool force)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
+#endif
